requirements.txt
requests==2.31.0
prometheus-client==0.20.0
pyyaml==6.0.2

